user_stack: compound_id bigint, matching compounds.id
The table was never applied to the current Supabase project, so stack
writes silently failed in prod (get_stack swallowed the missing-table
error and returned empty). Applied now with bigint per the trial_intakes
auditor convention.

+++ db/user_stack.sql
@@ -3,7 +3,8 @@
 create table if not exists user_stack (
   id            bigint generated always as identity primary key,
   user_ref      text not null,
-  compound_id   int  not null,
+  -- compounds.id is bigint identity; match it (see trial_intakes).
+  compound_id   bigint not null,
   compound_name text,
   dose          text,                  -- e.g. "7.5mg"
   source_type   text,                  -- label_dose | compounding_pharmacy | vendor_tested | gray_market
